Add reusable stubs fixture and update tests
